refactor(theme): include global reset manually
BREAKING CHANGES: Global styles must now be imported manually.
