+++ packages/aksara-ui-core/src/foundations/index.ts
@@ -3,5 +3,5 @@ export * from './card';
 export * from './common';
 export * from './focus-trap';
 export * from './portal';
-export * from './reset';
+export * from './theme';
 export * from './typography';
