+++ packages/aksara-ui-core/src/utils/storybook/SystemWrapper.tsx
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { DecoratorFn } from '@storybook/react';
 
-import { AksaraProvider } from '../../foundations/reset';
+import { AksaraProvider } from '../../foundations/theme';
 
 import '../../../../aksara-fonts/kata-icons.css';
 import '../../../../aksara-fonts/objective.css';
